requirements.txt
Flask
Pillow
pyiqa
torch
torchvision